src/requirements.txt
discord
aiosqlite
python-dotenv
pillow
psutil
